Caching prepared statements.

+++ Code/zhaw/components/src/main/java/ch/zhaw/ficore/p2abc/storage/SqliteURIBytesStorage.java
@@ -39,8 +39,14 @@ import com.google.inject.name.Named;
 public class SqliteURIBytesStorage extends URIBytesStorage {
 	private Connection con;
 	private String table;
-
+	
 	private Logger logger;
+    private PreparedStatement keysAsStringsStatement;
+    private PreparedStatement getStatement;
+    private PreparedStatement deleteStatement;
+    private PreparedStatement putStatement;
+    private PreparedStatement putNewStatement;
+    private PreparedStatement containsKeyStatement;
 
 	private static Map<String, Connection> connections = new HashMap<String, Connection>();
 
@@ -56,7 +62,7 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 	 */
 	@Inject
 	public SqliteURIBytesStorage(@Named("sqliteDBPath") String filePath, @Named("sqliteTblName") String table) throws ClassNotFoundException, SQLException, UnsafeTableNameException {
-		logger = LogManager.getLogger(SqliteURIBytesStorage.class.getName());
+		logger = LogManager.getLogger();
 		init(filePath, table);
 	}
 
@@ -91,6 +97,13 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 						" value         BLOB    NOT NULL)";
 				stmt.executeUpdate(sql);
 				this.table = table;
+				
+                keysAsStringsStatement = null;
+                getStatement = null;
+                deleteStatement = null;
+                putStatement = null;
+                putNewStatement = null;
+                containsKeyStatement = null;
 			}
 			catch(SQLException e) {
 				logger.catching(e);
@@ -150,13 +163,14 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 	public List<String> keysAsStrings() throws SQLException {
 		logger.entry();
 		synchronized(con) {
-			PreparedStatement pStmt = null;
 			ResultSet rst = null;
 			List<String> uris = new ArrayList<String>();
 
 			try {
-				pStmt = con.prepareStatement("SELECT uri FROM " + table);
-				rst = pStmt.executeQuery();
+			    if (keysAsStringsStatement == null)
+			        keysAsStringsStatement = con.prepareStatement("SELECT uri FROM " + table);
+			    
+				rst = keysAsStringsStatement.executeQuery();
 				while(rst.next()) {
 					try {
 						uris.add(rst.getString(1));
@@ -174,7 +188,7 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 				throw logger.throwing(new RuntimeException("Storage failure!"));
 			}
 			finally {
-				if(pStmt != null) pStmt.close();
+				if(keysAsStringsStatement != null) keysAsStringsStatement.close();
 				if(rst != null) rst.close();
 			}
 		}
@@ -187,14 +201,15 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 		logger.entry(key);
 		
 		synchronized(con) {
-			PreparedStatement pStmt = null;
 			ResultSet rst = null;
 
 			try {
-				pStmt = con.prepareStatement("SELECT value FROM " + table + " WHERE hash = ?");
+			    if (getStatement == null)
+			        getStatement = con.prepareStatement("SELECT value FROM " + table + " WHERE hash = ?");
+			    
 				String hash = DigestUtils.sha1Hex(key);
-				pStmt.setString(1, hash);
-				rst = pStmt.executeQuery();
+				getStatement.setString(1, hash);
+				rst = getStatement.executeQuery();
 				while(rst.next()) {
 				    byte[] ret = rst.getBytes(1);
 				    if (logger.isTraceEnabled()) {
@@ -209,7 +224,7 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 				throw logger.throwing(new RuntimeException("Storage failure!"));
 			}
 			finally {
-				if(pStmt != null) pStmt.close();
+				if(getStatement != null) getStatement.close();
 				if(rst != null) rst.close();
 			}
 		}
@@ -242,13 +257,12 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 	public void delete(String key) throws SQLException {
 		logger.entry();
 		synchronized(con) {
-			PreparedStatement pStmt = null;
-
 			try {
-				pStmt = con.prepareStatement("DELETE FROM " + table + " WHERE hash = ?");
+			    if (deleteStatement == null)
+			        deleteStatement = con.prepareStatement("DELETE FROM " + table + " WHERE hash = ?");
 				String hash = DigestUtils.sha1Hex(key);
-				pStmt.setString(1, hash);
-				pStmt.executeUpdate();
+				deleteStatement.setString(1, hash);
+				deleteStatement.executeUpdate();
 			}
 			catch(Exception e) {
 				logger.catching(e);
@@ -271,25 +285,25 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 
 			//Entry exists, so we need to do an UPDATE instead of an INSERT
 
-			PreparedStatement pStmt = null;
 			try {
-				pStmt = con.prepareStatement("UPDATE " + table + " SET uri = ?, value = ? WHERE " +
-						" hash = ?");
+			    if (putStatement == null)
+			        putStatement = con.prepareStatement("UPDATE " + table + " SET uri = ?, value = ? WHERE " +
+			                " hash = ?");
 
 				String hash = DigestUtils.sha1Hex(key);
 
-				pStmt.setString(3, hash);
-				pStmt.setString(1, key);
-				pStmt.setBytes(2, bytes);
+				putStatement.setString(3, hash);
+				putStatement.setString(1, key);
+				putStatement.setBytes(2, bytes);
 
-				pStmt.executeUpdate();
+				putStatement.executeUpdate();
 			}
 			catch(Exception e) {
 				logger.catching(e);
 				throw logger.throwing(new RuntimeException("Storage failure!"));
 			}
 			finally {
-				if(pStmt != null) pStmt.close();
+				if(putStatement != null) putStatement.close();
 			}
 		}
 		logger.exit();
@@ -304,18 +318,18 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 			if(containsKey(key))
 				return logger.exit(false);
 
-			PreparedStatement pStmt = null;
 			try {
-				pStmt = con.prepareStatement("INSERT INTO " + table + "(hash, uri, value) " +
-						"VALUES(?, ?, ?)");
+			    if (putNewStatement == null)
+			        putNewStatement = con.prepareStatement("INSERT INTO " + table + "(hash, uri, value) " +
+			                "VALUES(?, ?, ?)");
 
 				String hash = DigestUtils.sha1Hex(key);
 
-				pStmt.setString(1, hash);
-				pStmt.setString(2, key);
-				pStmt.setBytes(3, bytes);
+				putNewStatement.setString(1, hash);
+				putNewStatement.setString(2, key);
+				putNewStatement.setBytes(3, bytes);
 
-				pStmt.executeUpdate();
+				putNewStatement.executeUpdate();
 				return logger.exit(true);
 			}
 			catch(Exception e) {
@@ -323,7 +337,7 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 				throw logger.throwing(new RuntimeException("Storage failure!"));
 			}
 			finally {
-				if(pStmt != null) pStmt.close();
+				if(putNewStatement != null) putNewStatement.close();
 			}
 		}
 	}
@@ -334,14 +348,13 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 	public boolean containsKey(String key) throws SQLException {
 		logger.entry();
 		synchronized(con) {
-			PreparedStatement pStmt = null;
 			ResultSet rst = null;
 			try {
-				pStmt = con.prepareStatement("SELECT EXISTS(SELECT 1 FROM " + table + " WHERE " +
+				containsKeyStatement = con.prepareStatement("SELECT EXISTS(SELECT 1 FROM " + table + " WHERE " +
 						" hash = ? LIMIT 1)");
 				String hash = DigestUtils.sha1Hex(key);
-				pStmt.setString(1, hash);
-				rst = pStmt.executeQuery();
+				containsKeyStatement.setString(1, hash);
+				rst = containsKeyStatement.executeQuery();
 
 				if(!rst.next())
 					return logger.exit(false);
@@ -358,7 +371,7 @@ public class SqliteURIBytesStorage extends URIBytesStorage {
 				throw logger.throwing(new RuntimeException("Storage failure!"));
 			}
 			finally {
-				if(pStmt != null) pStmt.close();
+				if(containsKeyStatement != null) containsKeyStatement.close();
 				if(rst != null) rst.close();
 			}
 		}
